+++ networks/Model.py
@@ -1,0 +1,29 @@
+import torch
+import torch.nn as nn
+
+
+class Net(torch.nn.Module):
+
+    def __init__(self, dims):
+        super().__init__()
+        self.layers = []
+        for d in range(len(dims) - 1):
+            self.layers += [Layer(dims[d], dims[d + 1]).to(DEVICE)]
+
+    def predict(self, x):
+        goodness_per_label = []
+        for label in range(10):
+            h = overlay_y_on_x(x, label)
+            goodness = []
+            for layer in self.layers:
+                h = layer(h)
+                goodness += [h.pow(2).mean(1)]
+            goodness_per_label += [sum(goodness).unsqueeze(1)]
+        goodness_per_label = torch.cat(goodness_per_label, 1)
+        return goodness_per_label.argmax(1)
+
+    def train(self, x_pos, x_neg):
+        h_pos, h_neg = x_pos, x_neg
+        for i, layer in enumerate(self.layers):
+            print('training layer', i, '...')
+            h_pos, h_neg = layer.train(h_pos, h_neg)
